Refactor into config module
